better error handling for milestones
